Move the selector features directly into the Parser class
