+++ lib/index.ts
@@ -1,3 +1,2 @@
 export * from "./parser";
 export * from "./stream-parser";
-export * from "./selector";
